add: utils dir for momdularity, now metrics has its own script
