+++ stitch.py
@@ -33,7 +33,7 @@ import nibabel as nib
 from skimage.io import imread
 from skimage.transform import resize
 
-from utils import map_, tqdm_
+from utils.tensor_utils import map_, tqdm_
 
 
 def get_z(image: Path) -> int:
