+++ utils/tensor_utils.py
@@ -32,7 +32,7 @@ import torch
 import numpy as np
 from PIL import Image
 from tqdm import tqdm
-from torch import Tensor, einsum
+from torch import Tensor
 
 tqdm_ = partial(
     tqdm,
@@ -141,49 +141,3 @@ def save_images(segs: Tensor, names: Iterable[str], root: Path) -> None:
             np.save(str(save_path), seg.detach().cpu().numpy())
         else:
             raise ValueError(seg.shape)
-
-
-# Metrics
-def meta_dice(
-    sum_str: str, label: Tensor, pred: Tensor, smooth: float = 1e-8
-) -> Tensor:
-    assert label.shape == pred.shape
-    assert one_hot(label)
-    assert one_hot(pred)
-
-    inter_size: Tensor = einsum(sum_str, [intersection(label, pred)]).type(
-        torch.float32
-    )
-    sum_sizes: Tensor = (einsum(sum_str, [label]) + einsum(sum_str, [pred])).type(
-        torch.float32
-    )
-
-    dices: Tensor = (2 * inter_size + smooth) / (sum_sizes + smooth)
-
-    return dices
-
-
-dice_coef = partial(meta_dice, "bk...->bk")
-dice_batch = partial(meta_dice, "bk...->k")  # used for 3d dice
-
-
-def intersection(a: Tensor, b: Tensor) -> Tensor:
-    assert a.shape == b.shape
-    assert sset(a, [0, 1])
-    assert sset(b, [0, 1])
-
-    res = a & b
-    assert sset(res, [0, 1])
-
-    return res
-
-
-def union(a: Tensor, b: Tensor) -> Tensor:
-    assert a.shape == b.shape
-    assert sset(a, [0, 1])
-    assert sset(b, [0, 1])
-
-    res = a | b
-    assert sset(res, [0, 1])
-
-    return res
